backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
motor
pymongo
torch>=2.4.0
torchvision>=0.19.0
faiss-cpu
Pillow
python-dotenv
httpx
pytest
pytest-asyncio
numpy
opencv-python
invisible-watermark
PyWavelets
